sbi: Reference Resolution in specs complete (99%)
Added support for missing reference resolutions in `callbacks`,
`request_body` etc. A few references are not resolved yet, but instead
of chasing the rabit holes (for 1% unresolved references), we will fix
them as we incur issues during actual code generation. For now this
looks okay.

Using a fork of `openapi3` and pointing to 'git' for dependency
resolution. (Will create a separate crate if req. later).