sbi: Dependency respolution between specifications
Based on the passed files (The files we are interested in processing)
and any auxillary files (files to be loaded for resolving references),
we determine if there are any missing files and error out. Also, passing
a flag called `schema_only` that allows to only look for references
inside `/components/schemas`. This should allow us to process only those
references that are referred (and their dependent references -
transitively).