sbi: Code generation first pass
The reference resolution was still not complete. There are places, when
resolving references from the `aux_files` provided, we sometime have to
resolve references from within that `aux_file`. To be able to generate
struct definitions that are self contained, all such references have to
be resolved. Completed that resolution.

Also, basic code generation for struct and enum types started.