sbi: Initial fixes during code generation
When a Schema in `aux_files` refers back to one of the objects in the
given files, the back reference was incorrectly treated as a reference
that needs to be resolved, now if the 'reference' is in one of the
specs, we simply ignore that reference. This will eventually be
generated.