sbi: Fix issue with schema resolution in aux specs
Originally, the inner function
`get_component_schema_from_reference_in_spec` would try to resolve a
schema in an auxilary spec. However, such schema could have not only a
local reference (that is one that can be resolved from within the spec)
but also can have reference for something from one of the input specs.
If the local reference is unresolved, we need to check for the reference
in one of the input specs. This was not happening.

Also, avoided a nested if block by using the negative condition.