amf: `RanUe`s moved to `RanNode`
Previously `RanUe`s were maintained in the main `NgapManager` hashmap
with `gnb_ran_ue_id` as a key. However it is possible for two different
GNBs to use the same `gnb_ran_ue_id` key for different UEs. This would
result in conflicts. Moving this `HashMap` to `RanNode`, which is where
it should belong correctly.